third_party/WebKit/LayoutTests/fullscreen/api/document-fullscreen-element.html

<!DOCTYPE html>
<title>Document.fullscreenElement</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    var div = document.querySelector("div");

    document.onfullscreenchange = t.step_func(function()
    {
        assert_equals(document.fullscreenElement, div, "fullscreenElement before exitFullscreen()");
        document.exitFullscreen();
        // TODO(foolip): fullscreenElement should still be div.
        // https://crbug.com/402421
        assert_equals(document.fullscreenElement, null, "fullscreenElement after exitFullscreen()");

        document.onfullscreenchange = t.step_func(function()
        {
            assert_equals(document.fullscreenElement, null, "fullscreenElement after exiting fullscreen");
            t.done();
        });
    });

    trusted_click(t.step_func(function()
    {
        assert_equals(document.fullscreenElement, null, "fullscreenElement before requestFullscreen()");
        div.requestFullscreen();
        // TODO(foolip): fullscreenElement should still be null.
        // https://crbug.com/402421
        assert_equals(document.fullscreenElement, div, "fullscreenElement after requestFullscreen()");
    }), document.body);
});
</script>
